+++ src/modules/domain-links/index.ts
@@ -1,0 +1,4 @@
+export * from "@/modules/domain-links/interface";
+export * from "@/modules/domain-links/implementation";
+export * from "@/modules/domain-links/http";
+